typescript first version
