+++ src/resetModules.ts
@@ -1,9 +1,8 @@
-
 /*
 	utility to reset the state of any module 
 	(to it's default  state)
  */
-export const resetModules = dispatch => (modules = []) => {
+export const resetModules = (dispatch: any) => (modules: any[] = []) => {
   for (let i = 0; i < modules.length; i++) {
     let module = modules[i];
     dispatch({
